+++ scanner/Utilities.py
@@ -1,10 +1,6 @@
 # -*- coding: utf-8 -*-
 # do not remove previous line: it's not a comment!
 
-# @python2
-from __future__ import print_function
-from __future__ import division
-
 from datetime import datetime
 import os
 import json