+++ scanner/main.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
 import sys
@@ -16,10 +16,6 @@ from Utilities import report_times, message
 
 
 def main():
-	# @python2
-	if sys.version_info < (3,):
-		reload(sys)
-		sys.setdefaultencoding("UTF-8")
 	if len(sys.argv) != 3 and len(sys.argv) != 2:
 		print("usage: {0} ALBUM_PATH CACHE_PATH - or {1} CONFIG_FILE".format(sys.argv[0], sys.argv[0]))
 		return
@@ -30,7 +26,6 @@ def main():
 	from TreeWalker import TreeWalker
 
 	try:
-		# @python2
 		if sys.version_info >= (3, 4):
 			if Options.config['debug_memory']:
 				# Import the Debug module only if debugging is enabled
@@ -44,7 +39,6 @@ def main():
 		TreeWalker()
 		report_times(True)
 
-		# @python2
 		if sys.version_info >= (3, 4):
 			if Options.config['debug_profile']:
 				Debug.profile_stop()